modidiche clientcat, clientprod, controller e service

+++ NextCartFrontOffice_BE/src/main/java/it/nextre/nextcart/service/ListaSpesaServiceImpl.java
@@ -1,6 +1,6 @@
 package it.nextre.nextcart.service;
 
-import java.security.Principal;
+
 import java.util.List;
 import java.util.Optional;
 import org.jboss.logging.Logger;
@@ -14,6 +14,8 @@ import it.nextre.nextcart.entity.ListaSpesa;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 import jakarta.transaction.Transactional;
+import jakarta.ws.rs.ForbiddenException;
+import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.WebApplicationException;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.SecurityContext;
@@ -38,15 +40,19 @@ public class ListaSpesaServiceImpl implements ListaSpesaService {
 	public UserListaSpesaDTO createLista(ListaSpesaRequestDTO dto) {
 		
 		//Principal userPrincipal = securityContext.getUserPrincipal();
+		
 		//Long idUtente = Long.valueOf(userPrincipal.getName());
+		
 		Long idUtente = 23L;
 		
 		ListaSpesa listaSpesa = dto.toEntity();
 		
 		listaSpesa.setIdUtente(idUtente);
+		
 		repo.persist(listaSpesa);
 
 		UserListaSpesaDTO responseDTO = new UserListaSpesaDTO();
+		
 	    responseDTO.setListeSpesa(
 	        List.of(ListaSpesaSummaryDTO.fromEntity(listaSpesa))
 	    );
@@ -59,38 +65,56 @@ public class ListaSpesaServiceImpl implements ListaSpesaService {
 	@Transactional
 	public UserListaSpesaDTO getListeByUser() {
 		
-		// TODO: controllo se la lista non è vuota
         Long idUtente = 23L; 
 
 	    List<ListaSpesa> liste = repo.findByIdUtente(idUtente);
+	    
+        if (liste.isEmpty()) {
+            throw new WebApplicationException("Lista non trovata", Response.Status.NOT_FOUND);
+        }
+	    
 	    return UserListaSpesaDTO.fromEntity(idUtente, liste);
 	}
+	
 
 	@Override
 	@Transactional
 	public ListaSpesaResponseDTO getListaByIdAndUser(Long listaId) {
+		
         Long idUtente = 23L; 
 
         Optional<ListaSpesa> listaTrovata = repo.findByIdUtenteAndIdLista(listaId, idUtente);
+        
         if (listaTrovata.isEmpty()) {
             throw new WebApplicationException("Lista non trovata", Response.Status.NOT_FOUND);
         }
+        
         ListaSpesa lista = listaTrovata.get();
-        List<ProdottoDTO> prodottiShop = prodotti.getProdotti();
+        
+        List<ProdottoDTO> prodottiShop = prodotti.trovaTutti();
+        
         return ListaSpesaResponseDTO.fromEntity(lista, prodottiShop);
     }
 
 
 	@Override
 	@Transactional
-	public boolean deleteLista(Long listaId) {
-
-		// TODO: controllo se la lista è presente 
+	public boolean deleteLista(Long idUtente, Long listaId) {
 		
+        ListaSpesa trovata = repo.findById(listaId);
+        
+        if (trovata == null) {
+        	log.warn("Lista con id " + listaId + " non trovata.");
+            throw new NotFoundException("Lista associata al prodotto non trovata."); //TODO cambiare con eccezione custom
+        }
+        
+        if (!trovata.getIdUtente().equals(idUtente)) {
+        	log.warn("Utente con id: " + idUtente + "non autorizzato ad eliminare la lista con id: " + listaId);
+            throw new ForbiddenException("Non sei autorizzato a eliminare questa lista"); //TODO cambiare con eccezione custom
+        }
+ 
 		return repo.deleteById(listaId);
 		
 	}
 
-    
-
-    }
+   }
